refactor: move parsing of form submissions to form.ts
